requirements.txt
# ========================================
# ECG Monitor Application Dependencies
# ========================================

# Core GUI Framework
PyQt5>=5.15.0

# Scientific Computing and Signal Processing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# Real-time Plotting
pyqtgraph>=0.13.0

# Serial Communication (for ECG hardware)
pyserial>=3.5

# Data Processing and Analysis
pandas>=1.3.0

# Image Processing
Pillow>=8.3.0

# Audio Processing (for heartbeat sounds)
pyaudio>=0.2.11

# PDF Report Generation
reportlab>=3.6.0

# Cloud Storage (AWS S3)
boto3>=1.26.0

# Environment Configuration
python-dotenv>=0.19.0

# HTTP Requests (for API calls)
requests>=2.27.0

# Parsing Libraries
pyparsing>=3.0.0

# System Monitoring (for crash logger and diagnostics)
psutil>=5.9.0

# Security (for encrypted configuration)
cryptography>=41.0.0

# Optional: Enhanced Performance
numba>=0.56.0

# ========================================
# Standard Library Modules (Already Included in Python)
# ========================================
# The following are built-in Python modules and don't need installation:
# - json
# - csv
# - os
# - sys
# - time
# - datetime
# - threading
# - queue
# - socket
# - smtplib
# - email
# - traceback
# - logging
# - math
# - typing
# - pathlib
# - functools

# ========================================
# Development Dependencies (Optional)
# ========================================
# Uncomment these if you need development tools:
# pytest>=6.2.0          # Unit testing framework
# black>=21.0.0          # Code formatter
# flake8>=3.9.0          # Linter
# mypy>=0.910            # Type checker
# pytest-cov>=3.0.0      # Test coverage

# ========================================
# Notes for Installation:
# ========================================
# 1. Install all dependencies:
#    pip install -r requirements.txt
#
# 2. For macOS users (PyAudio requires PortAudio):
#    brew install portaudio
#    pip install pyaudio
#
# 3. For AWS S3 integration:
#    - Create a .env file with AWS credentials
#    - See .env.example for template
#
# 4. Virtual Environment (Recommended):
#    python3 -m venv .venv
#    source .venv/bin/activate  # macOS/Linux
#    .venv\Scripts\activate     # Windows
#    pip install -r requirements.txt
